# docs/swagger.yaml
basePath: /
definitions:
...
  organizations.CreateOrganizationDto:
    properties:
      description:
        type: string
      industry:
        type: string
      name:
        type: string
      owner_role:
        type: string
      primary_customer_type:
        enum:
        - b2b
        - b2c
        - both
        type: string
      team_size:
        type: string
      website_url:
        type: string
    required:
    - industry
    - name
    - owner_role
    - primary_customer_type
    - team_size
    type: object
  organizations.OrganizationResponseDto:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      industry:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      owner_role:
        type: string
      primary_customer_type:
        type: string
      team_size:
        type: string
      updated_at:
        type: string
      website_url:
        type: string
    type: object
  organizations.UpdateOrganizationDto:
    properties:
      description:
        type: string
      industry:
        type: string
      name:
        type: string
      owner_role:
        type: string
      primary_customer_type:
        enum:
        - b2b
        - b2c
        - both
        type: string
      team_size:
        type: string
      website_url:
        type: string
    required:
    - industry
    - name
    - owner_role
    - primary_customer_type
    - team_size
    type: object
...
  utils.EmptyData:
    type: object
  utils.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  utils.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  utils.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for Preuvio.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Preuvio API
  version: "1.0"
paths:
...
  /api/organizations:
    get:
      description: Retrieve a list of all organizations
      parameters:
      - description: Page (1-based, default 1)
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Limit (default 20, max 50)
        in: query
        maximum: 50
        minimum: 1
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/organizations.OrganizationResponseDto'
                  type: array
                meta:
                  $ref: '#/definitions/utils.PaginationMeta'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: List all organizations (paginated)
      tags:
      - organizations
    post:
      consumes:
      - application/json
      description: Creates a new organization and links it to the authenticated user.
        Requires Bearer token.
      parameters:
      - description: Organization data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/organizations.CreateOrganizationDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/organizations.OrganizationResponseDto'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new organization
      tags:
      - organizations
  /api/organizations/{id}:
    delete:
      description: Delete an organization by ID
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/utils.EmptyData'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Delete an organization
      tags:
      - organizations
    get:
      description: Retrieve organization details by its UUID
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/organizations.OrganizationResponseDto'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Get organization by ID
      tags:
      - organizations
    put:
      consumes:
      - application/json
      description: Update an existing organization by ID
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated organization data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/organizations.UpdateOrganizationDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/organizations.OrganizationResponseDto'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Update an organization
      tags:
      - organizations
  /api/organizations/me:
    get:
      description: Returns the organization the authenticated user belongs to. Requires
        Bearer token.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/organizations.OrganizationResponseDto'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get my organization
      tags:
      - organizations
...
swagger: "2.0"